Only show accrued traffic charges and estimated overages for the current month
For past months, customers can refer to their invoices instead. Hiding it when the selection is not the current month avoids weird things such as estimation errors due to to a month not having finished (vs what it actually *was* when it finished), potential changes in traffic package pricing, etc.