chore: use a local image of unleash enterprise for testing FE (#9509)
This removes the need of running a heroku instance with enterprise to
run frontend tests. We can also get rid of vercel deployments of the
frontend for PRs (ofc we'd be loosing the preview if we do this)

It relies on the free enterprise docker image to spin up a clean Unleash
instance on each frontend spec test.

It also seems to speed up tests, I guess because we don't have to wait
for vercel deployment queue:

![image](https://github.com/user-attachments/assets/60a2aef1-9a5a-442d-b241-8c5c145c679e)


Compared with other recent PRs:

![image](https://github.com/user-attachments/assets/9a7498aa-26e6-48a4-aba0-c67e5380ced4)

Unrelated to this PR but something is making the features spec tests
very slow since Cypress upgrade:
https://github.com/Unleash/unleash/actions/runs/13784760762/job/38550079932#step:4:138