Chore(1-3520)/playground disabled badges (#9583)
Gives a small update in how we deal with unevaluated and disabled
strategies in the new playground design:

- "Unevaluated" badges go from yellow warning to blue info and their
text changed to "Not evaluated"
- Don't show "Not evaluated" badges on strategies that are disabled.

To avoid this change affecting the current playground setup, I
duplicated the old resultschip into a legacy file and changed the
existing impl. To avoid updating all other files that use that chip
(it's all over the playground) and checking flags or creating duplicates
there, I decided to do a quick check at the top of the legacy file and
use the new file if the flag is on.

In doing so, I've also simplified the actual chip file and have more or
less cut the total line count in it in two 😄


![image](https://github.com/user-attachments/assets/44e38a8e-4faa-440b-82ed-9ee377160922)