feat(docs): Added ADR for logging levels (#9580)
This PR starts an ADR for logging levels, and how to make your
coworkers' days who are on-call less noisy.

This is my initial suggestion. Please feel free to help me with fleshing
out both the table for what log levels means to you or if you disagree
with what I've put as a background.

My impression is that we are pretty good at this, but we could be
better, and having an ADR to lean on when pushing back on PR's log
levels could help us.

Thanks to @kwasniew for the push to start the ADR.

Examples: 
*
https://github.com/Unleash/unleash/blob/038c10f6125c4cce200a0bf49f38c7bddada7093/src/lib/db/account-store.ts#L164-L173

And probably the storeFootprint method in our
client-feature-toggle-delta
*
https://github.com/Unleash/unleash/blob/038c10f6125c4cce200a0bf49f38c7bddada7093/src/lib/features/client-feature-toggles/delta/client-feature-toggle-delta.ts#L363-L370


*
https://github.com/bricks-software/unleash-enterprise/blob/293304a5d67231d584c3fa4c28589af23fb395e3/src/traffic-data/traffic-data-usage-service.ts#L69

---------

Co-authored-by: Melinda Fekete <melinda.fekete@getunleash.io>