remove mulitple constraints with react key test
The test was introduced in https://github.com/Unleash/unleash/pull/5917
and appears to relate to the autosave functionality. The autosave
functionality appears to be used primarily for undo and maybe for saving the state of a constraint before you "commit it". But both of these are things that we're removing, so I don't think there's any value in keeping this test around.